+++ http/src/util/futures.rs
@@ -3,7 +3,7 @@ pub(crate) use queue::*;
 
 #[cfg(any(feature = "http2", feature = "http3"))]
 mod queue {
-    use std::future::Future;
+    use core::future::Future;
 
     use futures_util::stream::{FuturesUnordered, StreamExt};
 
@@ -17,7 +17,7 @@ mod queue {
         #[cfg(any(all(feature = "http2", feature = "io-uring"), feature = "http3"))]
         pub(crate) async fn next(&mut self) -> F::Output {
             if self.is_empty() {
-                std::future::pending().await
+                core::future::pending().await
             } else {
                 self.next2().await
             }